requirements.txt
aiosqlite>=0.19.0
bcrypt>=4.1.0
PyJWT>=2.8.0
cryptography>=41.0.0
jieba>=0.42.1
pyyaml>=6.0
python-dotenv>=1.0.0
openai>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
